Reduce max width by 3 columns
This is necessary if we want to use this formatter with inquirer, 
because inquirer adds three extra columns of its own, which can cause
the longest stats to wrap to the next line in some cases.

Instead of trying to manipulate the stat after it has been generated,
it is much much easier to change the max width here, and let eslint-stats
do what it does best.